+++ ElvUI_EltreumUI/Modules/Skins/Borders.lua
@@ -164,10 +164,10 @@ function ElvUI_EltreumUI:Borders()
 				local isattachedtarget = E.db.unitframe.units.target.castbar.iconAttached
 				if isattachedtarget == false then
 					targetcastbarborder:SetSize(E.db.ElvUI_EltreumUI.borders.xcasttarget + E.db.unitframe.units.target.castbar.iconSize, E.db.ElvUI_EltreumUI.borders.ycasttarget)
-					targetcastbarborder:SetPoint("CENTER", _G["ElvUF_Target_CastBar"], "CENTER", -E.db.unitframe.units.target.castbar.iconSize/2, 0)
+					targetcastbarborder:SetPoint("CENTER", _G["ElvUF_Target_CastBar"], "CENTER", E.db.unitframe.units.target.castbar.iconSize/2, 0)
 				elseif isattachedtarget == true then
 					targetcastbarborder:SetSize(E.db.ElvUI_EltreumUI.borders.xcasttarget, E.db.ElvUI_EltreumUI.borders.ycasttarget)
-					targetcastbarborder:SetPoint("CENTER", _G["ElvUF_Target_CastBar"], "CENTER", -E.db.unitframe.units.target.castbar.iconSize/2, 0)
+					targetcastbarborder:SetPoint("CENTER", _G["ElvUF_Target_CastBar"], "CENTER", E.db.unitframe.units.target.castbar.iconSize/2, 0)
 				end
 				targetcastbarborder:SetBackdrop({
 					edgeFile = bordertexture,
@@ -682,6 +682,8 @@ function ElvUI_EltreumUI:AuraBorders(button)
 					edgeSize = 13,
 					})
 					auraborder:SetBackdropBorderColor(classcolor.r, classcolor.g, classcolor.b, 1)
+					auraborder:SetFrameStrata("MEDIUM")
+					auraborder:SetFrameLevel(4)
 				end
 			end
 		end