+++ ElvUI_EltreumUI/Options/OptionsRetail.lua
@@ -33,7 +33,7 @@ local THANKYOU = {
 	'|TInterface/AddOns/ElvUI_EltreumUI/Media/Textures/tukuidiscord:15:15:0:0:64:64:5:59:5:59|t Tukui Community for all the help and motivation',
 	'|TInterface/AddOns/ElvUI_EltreumUI/Media/Textures/addonsdiscord:15:15:0:0:64:64:5:59:5:59|t WoW AddOns; for answering a lot of questions',
 	'AcidWeb |TInterface/AddOns/ElvUI/Core/Media/ChatLogos/Gem:15:15:-1:2:64:64:6:60:8:60|t',
-	'|CFF8E44ADB|r|CFF2ECC71lin|r|CFF3498DBkii|r',
+	'|cff8E44ADB|r|cff2ECC71lin|r|cff3498DBkii|r',
 }
 
 local SUPPORTER = {
@@ -5448,7 +5448,7 @@ function ElvUI_EltreumUI:Configtable()
 								get = function() return E.db.ElvUI_EltreumUI.gradientmode.enableplayercastbar end,
 								set = function(_, value) E.db.ElvUI_EltreumUI.gradientmode.enableplayercastbar = value end,
 							},
-							enablecastbargradient = {
+							enablecastbargradientnoninterruptplayer = {
 								order = 2,
 								type = 'toggle',
 								name = L["Enable for Player Castbar (Non Interruptible)"],
@@ -5456,8 +5456,8 @@ function ElvUI_EltreumUI:Configtable()
 								width = 'full',
 								--disabled = function() return true end,
 								disabled = function() return not E.db.ElvUI_EltreumUI.gradientmode.enable end,
-								get = function() return E.db.ElvUI_EltreumUI.gradientmode.enableplayercastbar end,
-								set = function(_, value) E.db.ElvUI_EltreumUI.gradientmode.enableplayercastbar = value end,
+								get = function() return E.db.ElvUI_EltreumUI.gradientmode.enableplayercastbarnoninterruptible end,
+								set = function(_, value) E.db.ElvUI_EltreumUI.gradientmode.enableplayercastbarnoninterruptible = value end,
 							},
 							enabletargetcastbargradient = {
 								order = 2,
@@ -5469,6 +5469,17 @@ function ElvUI_EltreumUI:Configtable()
 								disabled = function() return not E.db.ElvUI_EltreumUI.gradientmode.enable end,
 								get = function() return E.db.ElvUI_EltreumUI.gradientmode.enabletargetcastbar end,
 								set = function(_, value) E.db.ElvUI_EltreumUI.gradientmode.enabletargetcastbar = value end,
+							},
+							enablecastbargradientnoninterrupttarget = {
+								order = 2,
+								type = 'toggle',
+								name = L["Enable for Target Castbar (Non Interruptible)"],
+								desc = L["Enable Gradient colors for Target Castbar (Non Interruptible)"],
+								width = 'full',
+								--disabled = function() return true end,
+								disabled = function() return not E.db.ElvUI_EltreumUI.gradientmode.enable end,
+								get = function() return E.db.ElvUI_EltreumUI.gradientmode.enabletargetcastbarnoninterruptible end,
+								set = function(_, value) E.db.ElvUI_EltreumUI.gradientmode.enabletargetcastbarnoninterruptible = value end,
 							},
 							enablecustomcolor = {
 								order = 3,
@@ -5508,7 +5519,7 @@ function ElvUI_EltreumUI:Configtable()
 							headercastbar = {
 								order = 4,
 								type = "description",
-								name = L["Castbar"],
+								name = L["Player Castbar"],
 								width = 'full',
 								image = function() return 'Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\EltreumHeader', 3240, 1 end,
 							},
@@ -5542,6 +5553,123 @@ function ElvUI_EltreumUI:Configtable()
 								end,
 								set = function(_, r, g, b, a)
 									E.db.ElvUI_EltreumUI.gradientmode.playercastbarR2, E.db.ElvUI_EltreumUI.gradientmode.playercastbarG2, E.db.ElvUI_EltreumUI.gradientmode.playercastbarB2 = r, g, b
+								end,
+							},
+							headercastbarcantinterruptplayer = {
+								order = 7,
+								type = "description",
+								name = L["Player Castbar (Non Interruptible)"],
+								width = 'full',
+								image = function() return 'Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\EltreumHeader', 3240, 1 end,
+							},
+							gradientcast1cantinterruptplayer = {
+								order = 8,
+								type = 'color',
+								name = L["Color 1"],
+								hasAlpha = false,
+								disabled = function() return not E.db.ElvUI_EltreumUI.gradientmode.enable or not E.db.ElvUI_EltreumUI.gradientmode.customcolor or not E.db.ElvUI_EltreumUI.gradientmode.enableplayercastbarnoninterruptible end,
+								get = function()
+									local dr = P.ElvUI_EltreumUI.gradientmode.playercastbarR1noninterruptible
+									local dg = P.ElvUI_EltreumUI.gradientmode.playercastbarG1noninterruptible
+									local db = P.ElvUI_EltreumUI.gradientmode.playercastbarB1noninterruptible
+									return E.db.ElvUI_EltreumUI.gradientmode.playercastbarR1noninterruptiblecustom, E.db.ElvUI_EltreumUI.gradientmode.playercastbarG1noninterruptiblecustom, E.db.ElvUI_EltreumUI.gradientmode.playercastbarB1noninterruptiblecustom, 1, dr, dg, db, 1
+								end,
+								set = function(_, r, g, b, a)
+									E.db.ElvUI_EltreumUI.gradientmode.playercastbarR1noninterruptiblecustom, E.db.ElvUI_EltreumUI.gradientmode.playercastbarG1noninterruptiblecustom, E.db.ElvUI_EltreumUI.gradientmode.playercastbarB1noninterruptiblecustom = r, g, b
+								end,
+							},
+							gradientcast2cantinterruptplayer = {
+								order = 9,
+								type = 'color',
+								name = L["Color 2"],
+								hasAlpha = false,
+								disabled = function() return not E.db.ElvUI_EltreumUI.gradientmode.enable or not E.db.ElvUI_EltreumUI.gradientmode.customcolor or not E.db.ElvUI_EltreumUI.gradientmode.enableplayercastbarnoninterruptible end,
+								get = function()
+									local dr = P.ElvUI_EltreumUI.gradientmode.playercastbarR2noninterruptible
+									local dg = P.ElvUI_EltreumUI.gradientmode.playercastbarG2noninterruptible
+									local db = P.ElvUI_EltreumUI.gradientmode.playercastbarB2noninterruptible
+									return E.db.ElvUI_EltreumUI.gradientmode.playercastbarR2noninterruptiblecustom, E.db.ElvUI_EltreumUI.gradientmode.playercastbarG2noninterruptiblecustom, E.db.ElvUI_EltreumUI.gradientmode.playercastbarB2noninterruptiblecustom, 1, dr, dg, db, 1
+								end,
+								set = function(_, r, g, b, a)
+									E.db.ElvUI_EltreumUI.gradientmode.playercastbarR2noninterruptiblecustom, E.db.ElvUI_EltreumUI.gradientmode.playercastbarG2noninterruptiblecustom, E.db.ElvUI_EltreumUI.gradientmode.playercastbarB2noninterruptiblecustom = r, g, b
+								end,
+							},
+							headercastbartarget = {
+								order = 10,
+								type = "description",
+								name = L["Target Castbar"],
+								width = 'full',
+								image = function() return 'Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\EltreumHeader', 3240, 1 end,
+							},
+							gradientcast1target = {
+								order = 11,
+								type = 'color',
+								name = L["Color 1"],
+								hasAlpha = false,
+								disabled = function() return not E.db.ElvUI_EltreumUI.gradientmode.enable or not E.db.ElvUI_EltreumUI.gradientmode.customcolor or not E.db.ElvUI_EltreumUI.gradientmode.enabletargetcastbar end,
+								get = function()
+									local dr = P.ElvUI_EltreumUI.gradientmode.targetcastbarR1
+									local dg = P.ElvUI_EltreumUI.gradientmode.targetcastbarG1
+									local db = P.ElvUI_EltreumUI.gradientmode.targetcastbarB1
+									return E.db.ElvUI_EltreumUI.gradientmode.targetcastbarR1custom, E.db.ElvUI_EltreumUI.gradientmode.targetcastbarG1custom, E.db.ElvUI_EltreumUI.gradientmode.targetcastbarB1custom, 1, dr, dg, db, 1
+								end,
+								set = function(_, r, g, b, a)
+									E.db.ElvUI_EltreumUI.gradientmode.targetcastbarR1custom, E.db.ElvUI_EltreumUI.gradientmode.targetcastbarG1custom, E.db.ElvUI_EltreumUI.gradientmode.targetcastbarB1custom = r, g, b
+								end,
+							},
+							gradientcast2target = {
+								order = 12,
+								type = 'color',
+								name = L["Color 2"],
+								hasAlpha = false,
+								disabled = function() return not E.db.ElvUI_EltreumUI.gradientmode.enable or not E.db.ElvUI_EltreumUI.gradientmode.customcolor or not E.db.ElvUI_EltreumUI.gradientmode.enabletargetcastbar end,
+								get = function()
+									local dr = P.ElvUI_EltreumUI.gradientmode.targetcastbarR2
+									local dg = P.ElvUI_EltreumUI.gradientmode.targetcastbarG2
+									local db = P.ElvUI_EltreumUI.gradientmode.targetcastbarB2
+									return E.db.ElvUI_EltreumUI.gradientmode.targetcastbarR2custom, E.db.ElvUI_EltreumUI.gradientmode.targetcastbarG2custom, E.db.ElvUI_EltreumUI.gradientmode.targetcastbarB2custom, 1, dr, dg, db, 1
+								end,
+								set = function(_, r, g, b, a)
+									E.db.ElvUI_EltreumUI.gradientmode.targetcastbarR2custom, E.db.ElvUI_EltreumUI.gradientmode.targetcastbarG2custom, E.db.ElvUI_EltreumUI.gradientmode.targetcastbarB2custom = r, g, b
+								end,
+							},
+							headercastbarnoninterruptibletargetcustom = {
+								order = 13,
+								type = "description",
+								name = L["Target Castbar (Non Interruptible)"],
+								width = 'full',
+								image = function() return 'Interface\\AddOns\\ElvUI_EltreumUI\\Media\\Textures\\EltreumHeader', 3240, 1 end,
+							},
+							gradientcast1noninterruptibletargetcustom = {
+								order = 14,
+								type = 'color',
+								name = L["Color 1"],
+								hasAlpha = false,
+								disabled = function() return not E.db.ElvUI_EltreumUI.gradientmode.enable or not E.db.ElvUI_EltreumUI.gradientmode.customcolor or not E.db.ElvUI_EltreumUI.gradientmode.enabletargetcastbarnoninterruptible end,
+								get = function()
+									local dr = P.ElvUI_EltreumUI.gradientmode.targetcastbarR1noninterruptible
+									local dg = P.ElvUI_EltreumUI.gradientmode.targetcastbarG1noninterruptible
+									local db = P.ElvUI_EltreumUI.gradientmode.targetcastbarB1noninterruptible
+									return E.db.ElvUI_EltreumUI.gradientmode.targetcastbarR1noninterruptiblecustom, E.db.ElvUI_EltreumUI.gradientmode.targetcastbarG1noninterruptiblecustom, E.db.ElvUI_EltreumUI.gradientmode.targetcastbarB1noninterruptiblecustom, 1, dr, dg, db, 1
+								end,
+								set = function(_, r, g, b, a)
+									E.db.ElvUI_EltreumUI.gradientmode.targetcastbarR1noninterruptiblecustom, E.db.ElvUI_EltreumUI.gradientmode.targetcastbarG1noninterruptiblecustom, E.db.ElvUI_EltreumUI.gradientmode.targetcastbarB1noninterruptiblecustom = r, g, b
+								end,
+							},
+							gradientcast2noninterruptibletargetcustom = {
+								order = 15,
+								type = 'color',
+								name = L["Color 2"],
+								hasAlpha = false,
+								disabled = function() return not E.db.ElvUI_EltreumUI.gradientmode.enable or not E.db.ElvUI_EltreumUI.gradientmode.customcolor or not E.db.ElvUI_EltreumUI.gradientmode.enabletargetcastbarnoninterruptible end,
+								get = function()
+									local dr = P.ElvUI_EltreumUI.gradientmode.targetcastbarR2noninterruptible
+									local dg = P.ElvUI_EltreumUI.gradientmode.targetcastbarG2noninterruptible
+									local db = P.ElvUI_EltreumUI.gradientmode.targetcastbarB2noninterruptible
+									return E.db.ElvUI_EltreumUI.gradientmode.targetcastbarR2noninterruptiblecustom, E.db.ElvUI_EltreumUI.gradientmode.targetcastbarG2noninterruptiblecustom, E.db.ElvUI_EltreumUI.gradientmode.targetcastbarB2noninterruptiblecustom, 1, dr, dg, db, 1
+								end,
+								set = function(_, r, g, b, a)
+									E.db.ElvUI_EltreumUI.gradientmode.targetcastbarR2noninterruptiblecustom, E.db.ElvUI_EltreumUI.gradientmode.targetcastbarG2noninterruptiblecustom, E.db.ElvUI_EltreumUI.gradientmode.targetcastbarB2noninterruptiblecustom = r, g, b
 								end,
 							},
 							header2 = {